Created utils directory and moved all sorting algorithms and animation utilities there
